requirements.txt
streamlit
numpy
pandas
scikit-learn
mecab-python3
pytorch-lightning
torch
torchmetrics
torchsummary